Move metric key logic to new file
